S15.f.3: extract plugin.ts chunked mesh transport (6890 → 6648 LOC)
Third sub-slice of S15.f plugin.ts decomposition. Lifts the
chunked-transport layer (>512 KB mesh payload split/reassemble) into
a dedicated module.

cli/src/core/mesh-transport.ts (new, ~265 LOC):
  - MESH_CHUNK_THRESHOLD/SIZE/MAX_CHUNKS/TRANSFER_TTL constants
  - PendingMeshTransfer interface
  - pendingTransfers Map + TTL cleanup setInterval
  - meshSendWithIdentity(client, target, message, identity, log)
  - meshHandleTransportMessage(fromAmid, fromAgent, message, log)

plugin.ts keeps a thin meshSend wrapper that captures agtIdentity, so
14+ call sites stay byte-identical. Signal Protocol crypto, X3DH, and
Double Ratchet remain in plugin.ts — agt-e2e-encryption skill's hard
rule preserved (this is JSON wire-shuffle only, not crypto).

Cumulative S15.f delta:
  pre:  7139
  f.1:  6974  (-165)
  f.2:  6890   (-84)
  f.3:  6648  (-242)
  Cumulative: -491. 5848 LOC remaining to §4.2 cap.

Verification:
  - tsc --noEmit clean
  - lint 29 warnings (+5 explicit-any disables on cross-boundary casts)
  - build clean
  - npm test: 454 pass / 2 skipped (same as baseline)

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f3.md
