+++ cli/src/plugin.ts
@@ -150,48 +150,7 @@ const AGT_RECONNECT_MAX_BACKOFF = 300_000; // 5 min cap
 // agent sends offload_task while the sandbox is already running the env task.
 const offloadInFlight = new Set<string>();
 
-// ── Chunked mesh transfer layer ──────────────────────────────────────────────
-// General-purpose transport for large payloads over the E2E encrypted mesh.
-// Any message exceeding MESH_CHUNK_THRESHOLD is auto-split into chunks by
-// meshSend(), and auto-reassembled by the onMessage handler before reaching
-// application logic. This is transparent — callers see a single send/receive.
-//
-// Protocol:
-//   1. Sender calls meshSend(amid, message)
-//   2. If serialized message ≤ threshold: sent as-is (fast path)
-//   3. If > threshold: sends mesh:transfer_manifest + N mesh:transfer_chunk
-//   4. Receiver accumulates chunks in pendingTransfers
-//   5. When all chunks arrive: verify SHA-256 hashes, reassemble, push to inbox
-//
-// Used by: mesh_send tool, mesh_transfer_file tool, handoff blob transfer,
-//          sub-agent workspace collection — any large mesh payload.
-
-const MESH_CHUNK_THRESHOLD = 512 * 1024;  // auto-chunk above 512KB
-const MESH_CHUNK_SIZE = 512 * 1024;       // 512KB per chunk
-const MESH_MAX_CHUNKS = 80;               // 80 × 512KB ≈ 40MB max payload
-const MESH_TRANSFER_TTL = 120_000;        // 2min TTL for incomplete transfers
-
-interface PendingMeshTransfer {
-  from_amid: string;
-  from_agent: string;
-  transfer_id: string;
-  total_chunks: number;
-  total_bytes: number;
-  chunk_hashes: string[];
-  manifest_hash: string;
-  metadata: Record<string, unknown>;  // all non-payload fields from the original message
-  chunks: Map<number, string>;
-  received_at: number;
-}
-const pendingTransfers = new Map<string, PendingMeshTransfer>();
-
-// Cleanup stale transfers every 30s
-setInterval(() => {
-  const now = Date.now();
-  for (const [key, t] of pendingTransfers) {
-    if (now - t.received_at > MESH_TRANSFER_TTL) pendingTransfers.delete(key);
-  }
-}, 30_000).unref();
+// Chunked mesh transport extracted to core/mesh-transport.ts in S15.f.3.
 
 // ── Handoff progress tracker (module-level, survives across tool calls) ──
 interface HandoffProgress {
@@ -345,6 +304,7 @@ async function notifyInboxToMemory(log: { info: (m: string) => void; warn: (m: s
 }
 import { discoverFoundryProject, type FoundryProjectInfo } from "./core/foundry-discovery.js";
 import { delegateToNativeAgent } from "./core/agt-task-delegate.js";
+import { meshSendWithIdentity, meshHandleTransportMessage, pendingTransfers, MESH_CHUNK_THRESHOLD, MESH_CHUNK_SIZE, MESH_MAX_CHUNKS, MESH_TRANSFER_TTL, type PendingMeshTransfer } from "./core/mesh-transport.js";
 let foundryProject: FoundryProjectInfo | null = null;
 let foundryInitialized = false;
 
@@ -1046,212 +1006,10 @@ async function meshSend(
   message: Record<string, unknown>,
   _log?: { info: (m: string) => void; warn: (m: string) => void },
 ): Promise<string | undefined> {
-  // Ed25519 per-message signing — attach signature and sender AMID
-  if (agtIdentity && !message.__signed) {
-    const payload = JSON.stringify(message);
-    try {
-      const encoder = new TextEncoder();
-      const signature = await agtIdentity.signB64(encoder.encode(payload));
-      (message as any).__signature = signature;
-      (message as any).__sender_amid = agtIdentity.amid;
-      (message as any).__signed_at = Math.floor(Date.now() / 1000);
-      // Push counter to router
-      pushSigningCounter("signed");
-    } catch {
-      // Sign failed — send unsigned (fail-open for availability)
-    }
-  }
-
-  const json = JSON.stringify(message);
-
-  if (json.length <= MESH_CHUNK_THRESHOLD) {
-    // Fast path — single message
-    await client.send(targetAmid, message);
-    return undefined;
-  }
-
-  // Large payload — chunked transfer
-  const totalChunks = Math.ceil(json.length / MESH_CHUNK_SIZE);
-  if (totalChunks > MESH_MAX_CHUNKS) {
-    throw new Error(
-      `Payload too large for mesh transfer: ${(json.length / 1024 / 1024).toFixed(1)} MB ` +
-      `(${totalChunks} chunks exceeds max ${MESH_MAX_CHUNKS})`
-    );
-  }
-
-  const { createHash } = await import("node:crypto");
-  const transferId = crypto.randomUUID();
-  const fromAgent = String(message.from_agent || process.env.SANDBOX_NAME || "unknown");
-
-  // Compute per-chunk SHA-256 for integrity verification
-  const chunkHashes: string[] = [];
-  for (let i = 0; i < totalChunks; i++) {
-    const chunk = json.slice(i * MESH_CHUNK_SIZE, (i + 1) * MESH_CHUNK_SIZE);
-    chunkHashes.push(createHash("sha256").update(chunk).digest("hex"));
-  }
-  const manifestHash = createHash("sha256").update(chunkHashes.join(":")).digest("hex");
-
-  _log?.info(
-    `Mesh chunked send: ${(json.length / 1024).toFixed(0)} KB → ${totalChunks} chunks ` +
-    `(transfer ${transferId.slice(0, 8)})`
-  );
-
-  // 1. Send manifest — receiver knows what to expect
-  await client.send(targetAmid, {
-    type: "mesh:transfer_manifest",
-    transfer_id: transferId,
-    original_type: message.type || "message",
-    total_chunks: totalChunks,
-    total_bytes: json.length,
-    chunk_hashes: chunkHashes,
-    manifest_hash: manifestHash,
-    from_agent: fromAgent,
-    timestamp: new Date().toISOString(),
-  });
-
-  // 2. Send chunks sequentially (relay preserves FIFO order per peer)
-  for (let i = 0; i < totalChunks; i++) {
-    const chunkData = json.slice(i * MESH_CHUNK_SIZE, (i + 1) * MESH_CHUNK_SIZE);
-    await client.send(targetAmid, {
-      type: "mesh:transfer_chunk",
-      transfer_id: transferId,
-      chunk_index: i,
-      total_chunks: totalChunks,
-      data: chunkData,
-      hash: chunkHashes[i],
-      from_agent: fromAgent,
-    });
-  }
-
-  _log?.info(`Mesh chunked send complete: ${totalChunks} chunks (transfer ${transferId.slice(0, 8)})`);
-  return transferId;
+  return meshSendWithIdentity(client, targetAmid, message, agtIdentity, _log);
 }
 
-// ── meshHandleTransportMessage: chunk accumulation + reassembly ───────────────
-// Called from onMessage handler. Returns the reassembled message if complete,
-// or null if the message was a transport chunk (absorbed, not for app layer).
-// Returns undefined if the message is NOT a transport message (pass through).
-async function meshHandleTransportMessage(
-  fromAmid: string,
-  fromAgent: string,
-  message: any,
-  _log?: { info: (m: string) => void; warn: (m: string) => void },
-): Promise<Record<string, unknown> | null | undefined> {
-  const msgType = message?.type;
-  if (msgType !== "mesh:transfer_manifest" && msgType !== "mesh:transfer_chunk") {
-    return undefined; // not a transport message — pass through to app layer
-  }
-
-  const transferId = message.transfer_id;
-  if (!transferId) return null; // malformed — absorb silently
-
-  const key = `${fromAmid}:${transferId}`;
-
-  if (msgType === "mesh:transfer_manifest") {
-    // Store manifest and prepare accumulator
-    pendingTransfers.set(key, {
-      from_amid: fromAmid,
-      from_agent: fromAgent,
-      transfer_id: transferId,
-      total_chunks: message.total_chunks,
-      total_bytes: message.total_bytes,
-      chunk_hashes: message.chunk_hashes || [],
-      manifest_hash: message.manifest_hash || "",
-      metadata: { original_type: message.original_type },
-      chunks: new Map(),
-      received_at: Date.now(),
-    });
-    _log?.info(
-      `Mesh transfer manifest: ${message.total_chunks} chunks, ` +
-      `${(message.total_bytes / 1024).toFixed(0)} KB (transfer ${transferId.slice(0, 8)})`
-    );
-    return null; // absorbed
-  }
-
-  // mesh:transfer_chunk
-  const transfer = pendingTransfers.get(key);
-  if (!transfer) {
-    // Chunk arrived before manifest or after TTL — try to find by transfer_id alone
-    for (const [k, t] of pendingTransfers) {
-      if (t.transfer_id === transferId && t.from_amid === fromAmid) {
-        t.chunks.set(message.chunk_index, message.data);
-        return null;
-      }
-    }
-    _log?.warn(`Mesh chunk for unknown transfer ${transferId.slice(0, 8)} — dropped`);
-    return null;
-  }
-
-  // Verify chunk hash before accepting
-  if (message.hash && transfer.chunk_hashes[message.chunk_index]) {
-    const { createHash } = await import("node:crypto");
-    const computed = createHash("sha256").update(message.data).digest("hex");
-    if (computed !== message.hash) {
-      _log?.warn(
-        `Mesh chunk ${message.chunk_index} hash mismatch (transfer ${transferId.slice(0, 8)}) — rejected`
-      );
-      return null; // reject corrupted chunk
-    }
-  }
-
-  transfer.chunks.set(message.chunk_index, message.data);
-
-  // Check if all chunks received
-  if (transfer.chunks.size < transfer.total_chunks) {
-    if (transfer.chunks.size % 10 === 0) {
-      _log?.info(
-        `Mesh transfer ${transferId.slice(0, 8)}: ${transfer.chunks.size}/${transfer.total_chunks} chunks`
-      );
-    }
-    return null; // still accumulating
-  }
-
-  // All chunks received — reassemble
-  _log?.info(
-    `Mesh transfer ${transferId.slice(0, 8)}: all ${transfer.total_chunks} chunks received — reassembling`
-  );
-
-  // Verify manifest hash (integrity of the complete chunk set)
-  const { createHash: mHash } = await import("node:crypto");
-  const actualHashes: string[] = [];
-  for (let i = 0; i < transfer.total_chunks; i++) {
-    actualHashes.push(
-      mHash("sha256").update(transfer.chunks.get(i) || "").digest("hex")
-    );
-  }
-  const actualManifestHash = mHash("sha256").update(actualHashes.join(":")).digest("hex");
-  if (transfer.manifest_hash && actualManifestHash !== transfer.manifest_hash) {
-    _log?.warn(
-      `Mesh transfer ${transferId.slice(0, 8)}: manifest hash mismatch — ` +
-      `data may be corrupted (expected ${transfer.manifest_hash.slice(0, 12)}, ` +
-      `got ${actualManifestHash.slice(0, 12)})`
-    );
-    // Continue anyway — let the application layer handle validation
-  }
-
-  // Reassemble JSON
-  const parts: string[] = [];
-  for (let i = 0; i < transfer.total_chunks; i++) {
-    parts.push(transfer.chunks.get(i) || "");
-  }
-  const reassembledJson = parts.join("");
-  pendingTransfers.delete(key);
-
-  let reassembled: Record<string, unknown>;
-  try {
-    reassembled = JSON.parse(reassembledJson);
-  } catch {
-    _log?.warn(`Mesh transfer ${transferId.slice(0, 8)}: reassembled JSON parse failed`);
-    return null;
-  }
-
-  _log?.info(
-    `Mesh transfer ${transferId.slice(0, 8)}: reassembled ${(reassembledJson.length / 1024).toFixed(0)} KB ` +
-    `(type: ${reassembled.type || transfer.metadata.original_type})`
-  );
-
-  return reassembled;
-}
+// meshHandleTransportMessage extracted to core/mesh-transport.ts in S15.f.3.
 
 // ── Offload task executor (shared between env-driven and message-driven flows) ──
 // Executes a task either via the native agent (delegateToNativeAgent) or the
